+++ activesupport/lib/active_support/core_ext/date_time/calculations.rb
@@ -208,4 +208,6 @@ class DateTime
       super
     end
   end
+  alias_method :before?, :<
+  alias_method :after?, :>
 end